refactor: modularise counselnotes and update test imports to match new counselnotes module structure
This commit updates the test imports to reflect the restructured counselnotes module, moving functions to specialized processors while maintaining compatibility with the existing API.
