+++ litassist/commands/counselnotes/__init__.py
@@ -1,0 +1,4 @@
+"""Counsel notes command package."""
+from .core import counselnotes
+
+__all__ = ["counselnotes"]
